… CMakePresets.json
… examples/CMakeLists.txt
… examples/baby-llama/baby-llama.cpp
… examples/benchmark/benchmark-matmult.cpp
… examples/cvector-generator/cvector-generator.cpp
… examples/export-lora/export-lora.cpp
… examples/llama-bench/llama-bench.cpp
… examples/llava/llava-cli.cpp
… examples/main/main.cpp
… examples/server/server.cpp
… examples/speculative/speculative.cpp
… ggml/CMakeLists.txt
… ggml/src/ggml-backend.c
… include/llama.h
… src/llama.cpp
… tests/test-rope.cpp