@fmz 2024-07-19
On W-2225 Xeon machine: CPU backend:

| CPU                                   | Model         | Test   |   t/s master |   t/s threadpool-attempt-2 |   Speedup |
|:--------------------------------------|:--------------|:-------|-------------:|---------------------------:|----------:|
| Intel(R) Xeon(R) W-2225 CPU @ 4.10GHz | llama 7B Q4_0 | pp512  |        17.57 |                      17.56 |      1.00 |
| Intel(R) Xeon(R) W-2225 CPU @ 4.10GHz | llama 7B Q4_0 | tg128  |         6.93 |                       7.14 |      1.03 |



$ ./build/bin/llama-bench -t 8 
| model                          |       size |     params | backend    | threads |          test |              t/s | 
| ------------------------------ | ---------: | ---------: | ---------- | ------: | ------------: | ---------------: |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CPU        |       8 |         pp512 |     18.38 ± 0.46 |
| llama 7B Q4_0                  |   3.56 GiB |     6.74 B | CPU        |       8 |         tg128 |      8.37 ± 0.01 |