@fmz 2024-07-22
./scripts/compare-commits.sh master threadpool-attempt-2 -t 1,2,4,6,8,10
| CPU                                       | Model         |   Threads | Test   |   t/s master |   t/s threadpool-attempt-2 |   Speedup |
|:------------------------------------------|:--------------|----------:|:-------|-------------:|---------------------------:|----------:|
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         1 | pp512  |         3.93 |                       3.94 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         1 | tg128  |         2.43 |                       2.44 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         2 | pp512  |         7.13 |                       7.06 |      0.99 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         2 | tg128  |         4.37 |                       4.36 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         4 | pp512  |        11.96 |                      11.99 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         4 | tg128  |         6.79 |                       6.77 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         6 | pp512  |        14.96 |                      14.98 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         6 | tg128  |         7.51 |                       7.53 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         8 | pp512  |        13.06 |                      13.09 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |         8 | tg128  |         6.88 |                       6.83 |      0.99 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |        10 | pp512  |        14.08 |                      14.06 |      1.00 |
| Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz | llama 7B Q4_0 |        10 | tg128  |         7.49 |                       7.52 |      1.00 |