@fmz 2024-07-22
$ LLAMA_CUDA=1 ./scripts/compare-commits.sh master threadpool-attempt-2 -nkvo 0,1
(No discernible difference between OpenMP and ggml_threadpool with default settings)

| GPU                 | Model         | NKVO   | Test   |   t/s master |   t/s threadpool-attempt-2 |   Speedup |
|:--------------------|:--------------|:-------|:-------|-------------:|---------------------------:|----------:|
| RTX 3060 Laptop GPU | llama 7B Q4_0 | No     | pp512  |      1644.73 |                    1642.34 |      1.00 |
| RTX 3060 Laptop GPU | llama 7B Q4_0 | No     | tg128  |        65.94 |                      65.89 |      1.00 |
| RTX 3060 Laptop GPU | llama 7B Q4_0 | Yes    | pp512  |       287.28 |                     286.44 |      1.00 |
| RTX 3060 Laptop GPU | llama 7B Q4_0 | Yes    | tg128  |        54.56 |                      54.32 |      1.00 |